0.2.04
Planetary Graphics updated
Moons added to normal sized worlds (gas giants coming)
Cities on planet's night side added according to planetary population
(might change to industrialised % of usable space)
Attempted storms but sadly they do not work for an unknown reason in builds but they do in play
Changed planetary population to be 1 = 1000 people
Struggled to try and balance economy, i think i need to re-work the generation of allegiances before i can finish the economy balancing

+++ Special Space Force/Assets/Scripts/Sector View/Planets/Planet_Class.cs
@@ -12,7 +12,7 @@ public class Planet_Class
     public string biome;
     public int biomeID;
     public bool inhabited;
-    public int population;
+    public float population;
     public int size;
     public float usableSpace;
     public float baseMetalsAmount;
